UI_Similarity/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
torch
transformers>=4.30.0
accelerate
torchvision